Can you guess what is the secret of academic success?
We have an excellent set which contains information about the life of students of mathematical courses and their academic success. As for me, I am very interested to examine in detail the information about the students and find out what influences their grades the most. The owner of the dataset suggests paying attention to the amount of alcohol consumed by students. This is a really interesting topic.